fix(ui): keep MCP header helpers browser-safe
